refactor filter component
